+++ bw.workflow.convictedreportinquirybysearchcriteria/src/main/java/sa/gov/nic/bio/bw/workflow/convictedreportinquirybysearchcriteria/utils/ConvictedReportInquiryErrorCodes.java
@@ -2,7 +2,7 @@ package sa.gov.nic.bio.bw.workflow.convictedreportinquirybysearchcriteria.utils;
 
 public enum ConvictedReportInquiryErrorCodes
 {
-	C014_00001, C014_00002, C014_00003, C014_00004, C014_00005, C014_00006, C014_00007, C014_00008, C014_00009;
+	C014_00001;
 	
 	public final String getCode()
 	{