+++ bw.workflow.registerconvictedpresent/src/main/java/sa/gov/nic/bio/bw/workflow/registerconvictedpresent/controllers/JudgmentDetailsPaneFxController.java
@@ -21,8 +21,8 @@ import sa.gov.nic.bio.bw.core.utils.AppUtils;
 import sa.gov.nic.bio.bw.core.utils.FxmlFile;
 import sa.gov.nic.bio.bw.core.utils.GuiUtils;
 import sa.gov.nic.bio.bw.core.workflow.Output;
-import sa.gov.nic.bio.bw.workflow.registerconvictedpresent.beans.CrimeType;
-import sa.gov.nic.bio.bw.workflow.registerconvictedpresent.lookups.CrimeTypesLookup;
+import sa.gov.nic.bio.bw.workflow.commons.beans.CrimeType;
+import sa.gov.nic.bio.bw.workflow.commons.lookups.CrimeTypesLookup;
 import sa.gov.nic.bio.bw.workflow.commons.beans.CrimeCode;
 
 import java.time.LocalDate;